backend/pyproject.toml
[tool.poetry]
name = "mahaguru-backend"
version = "0.1.0"
description = "Backend for Mahaguru AI platform"
authors = ["Mahaguru AI Team <dev@mahaguru.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.5.1"}
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.28.0"
weaviate-client = "^3.25.1"
openai = "^1.3.5"
langchain = "^0.0.350"
sentence-transformers = "^2.2.2"
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
types-requests = "^2.31.0.10"
types-python-dateutil = "^2.8.19.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["sqlalchemy.*", "alembic.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["jose.*"]
ignore_missing_imports = true
